packages/core/database/requirements.txt
# Core dependencies for database layer
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# PostgreSQL drivers
asyncpg==0.29.0        # Async driver для приложения
psycopg2-binary==2.9.9  # Sync driver для миграций

# Type hints and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
pdfplumber==0.11.0

# Development tools
mypy==1.8.0
black==24.1.1
ruff==0.1.13
